requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
Django==2.2.1
entrypoints==0.3
flake8==3.7.7
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Unipath==1.1
wcwidth==0.1.7
git+https://github.com/ptronico/django-nungoforms.git@0.1#django-nungoforms
